// packages/test/generatedCode/users/item/outlook/taskGroups/item/outlookTaskGroupItemRequestBuilderGetRequestConfiguration.ts
import {OutlookTaskGroupItemRequestBuilderGetQueryParameters} from './outlookTaskGroupItemRequestBuilderGetQueryParameters';
import {RequestOption} from '@microsoft/kiota-abstractions';

export interface OutlookTaskGroupItemRequestBuilderGetRequestConfiguration {
    /** Request headers */
    headers?: Record<string, string>;
    /** Request options */
    options?: RequestOption[];
    /** Request query parameters */
    queryParameters?: OutlookTaskGroupItemRequestBuilderGetQueryParameters;
}